requirements.txt
onnxruntime
transformers
pillow
numpy
python-dotenv
tqdm
requests
